Add files via upload
CRUD stands for Create, Read (or retrieve), Update, Delete. It is a Java Server Page (JSP) project with a JavaBean that is used to perfom the create read update and delete operations. For the database it uses java derby 'jdbc:derby://localhost:1527/sample' and table STUDENTS with 3 columns namely First name, Last name and ID number. I've used Semantic UI and w3.css for styling and JQuery.